리뷰적용 - event 함수들 RacingViewModule로 이동
